Relocate playwright cloud runner
